requirements.txt
diff-match-patch==20121119
Django==2.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-import-export==0.6.1
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
mysqlclient==1.3.12
odfpy==1.3.6
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
pytz==2017.3
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
